Moved `toURL` to `helpers/uri`
no refs

Moved the `toURL` helper to the URI helpers and updated all references to it
