+++ src/handlers.ts
@@ -16,7 +16,6 @@ import {
 import { Context, Next } from 'hono';
 import { v4 as uuidv4 } from 'uuid';
 import { addToList, removeFromList } from './kv-helpers';
-import { toURL } from './toURL';
 import { ContextData, HonoContextVariables, fedify } from './app';
 import { getSiteSettings } from './helpers/ghost';
 import type { PersonData } from './user';
@@ -24,6 +23,7 @@ import { ACTOR_DEFAULT_HANDLE } from './constants';
 import { Temporal } from '@js-temporal/polyfill';
 import { createHash } from 'node:crypto';
 import { lookupActor } from 'lookup-helpers';
+import { toURL } from './helpers/uri';
 import { buildActivity } from 'helpers/activitypub/activity';
 
 import z from 'zod';
