+++ src/helpers/uri.ts
@@ -7,3 +7,17 @@ export function isUri(value: string): boolean {
         return false;
     }
 }
+
+export function toURL(value: unknown): URL | undefined {
+    if (value instanceof URL) {
+        return value;
+    }
+    if (typeof value !== 'string') {
+        return undefined;
+    }
+    try {
+        return new URL(value);
+    } catch (err) {
+        return undefined;
+    }
+}
